finalize the wrapper and restructure the code
